'Flatten components folder for future refactoring'
